refactor: split runner, checker into separate modules
